# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TopRopes Backend API
  version: 0.2.0-draft
  x-summary: REST contract used by the TopRopes frontend.
  description: |
    This document is the REST contract for ring-rumble-react and the Spring implementation.

    Current UI identity model:
    - Users enter username + password only.
    - The backend issues JWT access and refresh tokens.

    Current UI authorization model:
    - Catalogue data is publicly readable.
    - User-owned ratings, dossiers, and promos require authentication.
    - Catalogue mutations require an administrator role.
  x-contract-status: implemented
  x-current-frontend-runtime: Spring Boot JWT API and PostgreSQL
servers:
  - url: https://api.topropes.example.com/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /users/me/ratings:
    get:
      tags: [Ratings]
      summary: List current user's ratings
      operationId: listMyRatings
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Ratings page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchRatingPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/ratings/{matchSlug}:
    put:
      tags: [Ratings]
      summary: Create or update rating for a match
      operationId: upsertMyRating
      security:
        - bearerAuth: []
      parameters:
        - name: matchSlug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchRatingUpsertRequest'
      responses:
        '200':
          description: Rating persisted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchRating'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Ratings]
      summary: Delete rating for a match
      operationId: deleteMyRating
      security:
        - bearerAuth: []
      parameters:
        - name: matchSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Rating removed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/me/feuds/{feudSlug}/dossier:
    get:
      tags: [FeudDossiers]
      summary: Get current user's dossier for a feud
      operationId: getMyFeudDossier
      security:
        - bearerAuth: []
      parameters:
        - name: feudSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dossier result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeudDossier'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [FeudDossiers]
      summary: Create or update current user's dossier for a feud
      operationId: upsertMyFeudDossier
      security:
        - bearerAuth: []
      parameters:
        - name: feudSlug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeudDossierUpsertRequest'
      responses:
        '200':
          description: Dossier persisted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeudDossier'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [FeudDossiers]
      summary: Delete current user's dossier for a feud
      operationId: deleteMyFeudDossier
      security:
        - bearerAuth: []
      parameters:
        - name: feudSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Dossier removed.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/feuds/{feudSlug}/promos:
    get:
      tags: [FeudPromos]
      summary: List current user's promos for a feud
      operationId: listMyFeudPromos
      security:
        - bearerAuth: []
      parameters:
        - name: feudSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Promos list.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/FeudPromo'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/feuds/{feudSlug}/promos/{promoSlug}:
    put:
      tags: [FeudPromos]
      summary: Create or update current user's feud promo
      operationId: upsertMyFeudPromo
      security:
        - bearerAuth: []
      parameters:
        - name: feudSlug
          in: path
          required: true
          schema:
            type: string
        - name: promoSlug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeudPromoUpsertRequest'
      responses:
        '200':
          description: Promo persisted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeudPromo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [FeudPromos]
      summary: Delete current user's feud promo
      operationId: deleteMyFeudPromo
      security:
        - bearerAuth: []
      parameters:
        - name: feudSlug
          in: path
          required: true
          schema:
            type: string
        - name: promoSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Promo removed.
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationError:
      description: Request payload or query values failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: array
              items:
                $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
...
    MatchRating:
      type: object
      required:
        [matchSlug, crowd, story, difficulty, technique, personalStars, review, notes, updatedAt]
      properties:
        matchSlug:
          type: string
        crowd:
          type: integer
          minimum: 0
          maximum: 10
        story:
          type: integer
          minimum: 0
          maximum: 10
        difficulty:
          type: integer
          minimum: 0
          maximum: 10
        technique:
          type: integer
          minimum: 0
          maximum: 10
        personalStars:
          type: number
          minimum: 0
          maximum: 5
          multipleOf: 0.5
        review:
          type: string
        notes:
          type: object
          additionalProperties:
            type: string
        updatedAt:
          type: string
          format: date-time
    MatchRatingUpsertRequest:
      type: object
      required: [crowd, story, difficulty, technique, personalStars, review, notes]
      properties:
        crowd:
          type: integer
          minimum: 0
          maximum: 10
        story:
          type: integer
          minimum: 0
          maximum: 10
        difficulty:
          type: integer
          minimum: 0
          maximum: 10
        technique:
          type: integer
          minimum: 0
          maximum: 10
        personalStars:
          type: number
          minimum: 0
          maximum: 5
          multipleOf: 0.5
        review:
          type: string
        notes:
          type: object
          additionalProperties:
            type: string
    KeyDevelopment:
      type: object
      required: [date, label, detail, heatDelta]
      properties:
        date:
          type: string
        label:
          type: string
        detail:
          type: string
        heatDelta:
          type: integer
    FeudDossier:
      type: object
      required:
        [feudSlug, participants, cause, motivationA, motivationB, background, stakes, currentStatus, resolution, keyDevelopments, heat, updatedAt]
      properties:
        feudSlug:
          type: string
        participants:
          type: string
        cause:
          type: string
        motivationA:
          type: string
        motivationB:
          type: string
        background:
          type: string
        stakes:
          type: string
        currentStatus:
          type: string
        resolution:
          type: string
        keyDevelopments:
          type: array
          items:
            $ref: '#/components/schemas/KeyDevelopment'
        heat:
          type: integer
          minimum: 0
          maximum: 100
        updatedAt:
          type: string
          format: date-time
    FeudDossierUpsertRequest:
      type: object
      required:
        [participants, cause, motivationA, motivationB, background, stakes, currentStatus, resolution, keyDevelopments, heat]
      properties:
        participants:
          type: string
        cause:
          type: string
        motivationA:
          type: string
        motivationB:
          type: string
        background:
          type: string
        stakes:
          type: string
        currentStatus:
          type: string
        resolution:
          type: string
        keyDevelopments:
          type: array
          items:
            $ref: '#/components/schemas/KeyDevelopment'
        heat:
          type: integer
          minimum: 0
          maximum: 100
    FeudPromo:
      type: object
      required: [feudSlug, promoSlug, title, speaker, date, venue, quote, transcript, impact, updatedAt]
      properties:
        feudSlug:
          type: string
        promoSlug:
          type: string
        title:
          type: string
        speaker:
          type: string
        date:
          type: string
          format: date
        venue:
          type: string
        quote:
          type: string
        transcript:
          type: string
        impact:
          type: string
        updatedAt:
          type: string
          format: date-time
    FeudPromoUpsertRequest:
      type: object
      required: [title, speaker, date, venue, quote, transcript, impact]
      properties:
        title:
          type: string
        speaker:
          type: string
        date:
          type: string
          format: date
        venue:
          type: string
        quote:
          type: string
        transcript:
          type: string
        impact:
          type: string
...
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
...
    MatchRatingPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MatchRating'
        meta:
          $ref: '#/components/schemas/PageMeta'